# specs/001-aegisai-mvp-foundation/contracts/mvp-openapi.yaml
openapi: 3.1.0
info:
  title: AegisAI MVP API
  version: 0.1.0
  description: Baseline contract for the AegisAI MVP defined in spec 2.2.md.
servers:
  - url: http://localhost:3000/api
security:
  - sessionCookie: []
paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      security: []
      responses:
        '200':
          description: Raw health payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /auth/github:
    get:
      summary: Start GitHub OAuth login
      operationId: githubLogin
      security: []
      responses:
        '302':
          description: Redirect to GitHub OAuth consent
  /auth/github/callback:
    get:
      summary: Complete GitHub OAuth login
      operationId: githubCallback
      security: []
      responses:
        '302':
          description: Redirect to the frontend dashboard
  /auth/gitlab:
    get:
      summary: Start GitLab OAuth login
      operationId: gitlabLogin
      security: []
      responses:
        '302':
          description: Redirect to GitLab OAuth consent
  /auth/gitlab/callback:
    get:
      summary: Complete GitLab OAuth login
      operationId: gitlabCallback
      security: []
      responses:
        '302':
          description: Redirect to the frontend dashboard
  /auth/me:
    get:
      summary: Get the current authenticated user
      operationId: getCurrentUser
      responses:
        '200':
          description: Authenticated user response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      summary: Log out the current user
      operationId: logout
      responses:
        '200':
          description: Session cleared
          content:
            application/json:
              schema:
                type: 'null'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /repos:
    get:
      summary: List connected repositories
      operationId: listConnectedRepos
      responses:
        '200':
          description: Connected repositories for the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectedRepoListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Connect a repository
      operationId: connectRepo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectRepoRequest'
      responses:
        '201':
          description: Repository connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectedRepoResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /repos/available:
    get:
      summary: List repositories available from connected providers
      operationId: listAvailableRepos
      parameters:
        - name: provider
          in: query
          required: true
          schema:
            type: string
            enum: [github, gitlab]
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: size
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Paginated provider repository list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderRepoPageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /repos/{repoId}:
    delete:
      summary: Disconnect a repository
      operationId: disconnectRepo
      parameters:
        - $ref: '#/components/parameters/RepoId'
      responses:
        '204':
          description: Repository disconnected
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /repos/{repoId}/branches:
    get:
      summary: List branches for a connected repository
      operationId: listBranches
      parameters:
        - $ref: '#/components/parameters/RepoId'
      responses:
        '200':
          description: Branch list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchPageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /repos/{repoId}/scans:
    get:
      summary: List scans for a connected repository
      operationId: listRepoScans
      parameters:
        - $ref: '#/components/parameters/RepoId'
      responses:
        '200':
          description: Scan history page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepoScanPageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /scans:
    post:
      summary: Request a repository scan
      operationId: requestScan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestScanRequest'
      responses:
        '202':
          description: Scan accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanRequestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /scans/{scanId}:
    get:
      summary: Get scan status and summary
      operationId: getScan
      parameters:
        - $ref: '#/components/parameters/ScanId'
      responses:
        '200':
          description: Scan detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /scans/{scanId}/vulnerabilities:
    get:
      summary: List vulnerabilities for a scan
      operationId: listScanVulnerabilities
      parameters:
        - $ref: '#/components/parameters/ScanId'
      responses:
        '200':
          description: Vulnerability page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VulnerabilityPageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '404':
          $ref: '#/components/responses/NotFound'
  /vulnerabilities/{vulnId}:
    get:
      summary: Get vulnerability detail
      operationId: getVulnerability
      parameters:
        - $ref: '#/components/parameters/VulnId'
      responses:
        '200':
          description: Vulnerability detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VulnerabilityDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '404':
          $ref: '#/components/responses/NotFound'
  /vulnerabilities/{vulnId}/feedback:
    post:
      summary: Submit user feedback for a vulnerability
      operationId: submitVulnerabilityFeedback
      parameters:
        - $ref: '#/components/parameters/VulnId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VulnerabilityFeedbackRequest'
      responses:
        '200':
          description: Recorded reviewer feedback
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VulnerabilityFeedbackResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /dashboard:
    get:
      summary: Get dashboard summary
      operationId: getDashboard
      responses:
        '200':
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /reports/scans/{scanId}/pdf:
    post:
      summary: Request PDF report generation for a completed scan
      operationId: requestPdfReport
      parameters:
        - $ref: '#/components/parameters/ScanId'
      responses:
        '202':
          description: Report generation accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportRequestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /reports/{reportId}:
    get:
      summary: Get report status
      operationId: getReport
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Report detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /reports/{reportId}/download:
    get:
      summary: Download a generated PDF report
      operationId: downloadReport
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: PDF binary
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: Report expired
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
  parameters:
    RepoId:
      name: repoId
      in: path
      required: true
      schema:
        type: string
    ScanId:
      name: scanId
      in: path
      required: true
      schema:
        type: string
    VulnId:
      name: vulnId
      in: path
      required: true
      schema:
        type: string
    ReportId:
      name: reportId
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
    Unauthorized:
      description: Unauthorized
    Forbidden:
      description: Forbidden
    NotFound:
      description: Resource not found
    Conflict:
      description: Conflict
    TooManyRequests:
      description: Too many requests
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
        uptime:
          type: number
        services:
          type: object
          additionalProperties:
            type: string
        timestamp:
          type: string
      required: [status, uptime, services, timestamp]
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
        name:
          type: string
        avatarUrl:
          type: [string, 'null']
        connectedProviders:
          type: array
          items:
            type: string
            enum: [github, gitlab]
      required: [id, name, connectedProviders]
    ProviderRepo:
      type: object
      properties:
        providerRepoId:
          type: string
        fullName:
          type: string
        cloneUrl:
          type: string
        defaultBranch:
          type: string
        isPrivate:
          type: boolean
        alreadyConnected:
          type: boolean
      required: [providerRepoId, fullName, cloneUrl, defaultBranch, isPrivate, alreadyConnected]
    ConnectedRepo:
      type: object
      properties:
        id:
          type: string
        provider:
          type: string
          enum: [github, gitlab]
        fullName:
          type: string
        cloneUrl:
          type: string
        defaultBranch:
          type: string
        isPrivate:
          type: boolean
        lastScanAt:
          type: [string, 'null']
        lastScanStatus:
          type: [string, 'null']
          enum: [PENDING, RUNNING, DONE, FAILED, null]
      required: [id, provider, fullName, cloneUrl, defaultBranch, isPrivate, lastScanAt, lastScanStatus]
    Branch:
      type: object
      properties:
        name:
          type: string
        isDefault:
          type: boolean
        lastCommitSha:
          type: [string, 'null']
      required: [name, isDefault, lastCommitSha]
    ScanSummary:
      type: object
      properties:
        critical:
          type: integer
        high:
          type: integer
        medium:
          type: integer
        low:
          type: integer
        info:
          type: integer
      required: [critical, high, medium, low, info]
    ScanDetail:
      type: object
      properties:
        id:
          type: string
        repoFullName:
          type: string
        branch:
          type: string
        commitSha:
          type: [string, 'null']
        status:
          type: string
          enum: [PENDING, RUNNING, DONE, FAILED]
        language:
          type: string
        totalFiles:
          type: [integer, 'null']
        totalLines:
          type: [integer, 'null']
        summary:
          $ref: '#/components/schemas/ScanSummary'
        startedAt:
          type: [string, 'null']
        completedAt:
          type: [string, 'null']
        errorMessage:
          type: [string, 'null']
        createdAt:
          type: string
      required: [id, repoFullName, branch, commitSha, status, language, totalFiles, totalLines, summary, startedAt, completedAt, errorMessage, createdAt]
    ModelResult:
      type: object
      properties:
        model:
          type: string
        detected:
          type: boolean
        severity:
          type: string
        reasoning:
          type: string
      required: [model, detected, severity, reasoning]
    ReferenceLink:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
      required: [title, url]
    Vulnerability:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        severity:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW, INFO]
        filePath:
          type: string
        lineStart:
          type: integer
        lineEnd:
          type: [integer, 'null']
        cweId:
          type: [string, 'null']
        owaspCategory:
          type: [string, 'null']
        status:
          type: string
          enum: [OPEN, FIXED, ACCEPTED, REJECTED]
      required: [id, title, severity, filePath, lineStart, status]
    VulnerabilityDetail:
      allOf:
        - $ref: '#/components/schemas/Vulnerability'
        - type: object
          properties:
            description:
              type: string
            fixSuggestion:
              type: [string, 'null']
            fixExplanation:
              type: [string, 'null']
            codeSnippet:
              type: [string, 'null']
            cveId:
              type: [string, 'null']
            referenceLinks:
              type: [array, 'null']
              items:
                $ref: '#/components/schemas/ReferenceLink'
            consensusScore:
              type: [number, 'null']
            modelResults:
              type: [array, 'null']
              items:
                $ref: '#/components/schemas/ModelResult'
            userFeedback:
              type: [string, 'null']
              enum: [ACCEPTED, REJECTED, null]
          required: [description]
    TrendItem:
      type: object
      properties:
        date:
          type: string
        critical:
          type: integer
        high:
          type: integer
        medium:
          type: integer
      required: [date, critical, high, medium]
    DashboardWorkspace:
      type: object
      properties:
        totalRepos:
          type: integer
        totalScans:
          type: integer
        openVulnerabilities:
          $ref: '#/components/schemas/ScanSummary'
        connectedRepos:
          type: array
          items:
            $ref: '#/components/schemas/ConnectedRepo'
        completedScans:
          type: array
          items:
            $ref: '#/components/schemas/ScanDetail'
        severitySummary:
          $ref: '#/components/schemas/ScanSummary'
        degraded:
          type: boolean
        trend:
          type: array
          items:
            $ref: '#/components/schemas/TrendItem'
        recentScans:
          type: array
          items:
            $ref: '#/components/schemas/ScanDetail'
      required:
        [
          totalRepos,
          totalScans,
          openVulnerabilities,
          connectedRepos,
          completedScans,
          severitySummary,
          degraded,
          trend,
          recentScans
        ]
    ReportDetail:
      type: object
      properties:
        id:
          type: string
        scanId:
          type: string
        status:
          type: string
          enum: [GENERATING, READY, FAILED, EXPIRED]
        downloadUrl:
          type: [string, 'null']
        errorMessage:
          type: [string, 'null']
        createdAt:
          type: string
        expiresAt:
          type: [string, 'null']
      required: [id, scanId, status, downloadUrl, errorMessage, createdAt, expiresAt]
    ApiEnvelopeAuthUser:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/AuthUser'
      required: [success, data]
    ApiEnvelopeConnectedRepo:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/ConnectedRepo'
      required: [success, data]
    ApiEnvelopeScanRequest:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/ScanRequestAccepted'
      required: [success, data]
    ApiEnvelopeScanDetail:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/ScanDetail'
      required: [success, data]
    ApiEnvelopeVulnerabilityDetail:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/VulnerabilityDetail'
        message:
          type: [string, 'null']
        timestamp:
          type: string
      required: [success, data, message, timestamp]
    ApiEnvelopeVulnerabilityFeedback:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/VulnerabilityFeedback'
        message:
          type: [string, 'null']
        timestamp:
          type: string
      required: [success, data, message, timestamp]
    ApiEnvelopeDashboard:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/DashboardWorkspace'
      required: [success, data]
    ApiEnvelopeReportDetail:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/ReportDetail'
        message:
          type: [string, 'null']
        timestamp:
          type: string
      required: [success, data, message, timestamp]
    ProviderRepoPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProviderRepo'
        page:
          type: integer
        size:
          type: integer
        hasNextPage:
          type: boolean
        nextPage:
          type: [integer, 'null']
      required: [items, page, size, hasNextPage, nextPage]
    BranchPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Branch'
        page:
          type: integer
        size:
          type: integer
        hasNextPage:
          type: boolean
        nextPage:
          type: [integer, 'null']
      required: [items, page, size, hasNextPage, nextPage]
    ConnectedRepoList:
      type: array
      items:
        $ref: '#/components/schemas/ConnectedRepo'
    RepoScanPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ScanDetail'
        totalCount:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [items, totalCount, page, totalPages]
    VulnerabilityPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Vulnerability'
        page:
          type: integer
        totalCount:
          type: integer
        totalPages:
          type: integer
      required: [items, page, totalCount, totalPages]
    ProviderRepoPageResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/ProviderRepoPage'
        message:
          type: [string, 'null']
        timestamp:
          type: string
      required: [success, data, message, timestamp]
    BranchPageResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/BranchPage'
        message:
          type: [string, 'null']
        timestamp:
          type: string
      required: [success, data, message, timestamp]
    VulnerabilityPageResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/VulnerabilityPage'
        message:
          type: [string, 'null']
        timestamp:
          type: string
      required: [success, data, message, timestamp]
    ConnectRepoRequest:
      type: object
      properties:
        provider:
          type: string
          enum: [github, gitlab]
        providerRepoId:
          type: string
      required: [provider, providerRepoId]
    RequestScanRequest:
      type: object
      properties:
        repoId:
          type: string
        branch:
          type: string
      required: [repoId, branch]
    ScanRequestAccepted:
      type: object
      properties:
        scanId:
          type: string
        status:
          type: string
          enum: [PENDING]
      required: [scanId, status]
    VulnerabilityFeedbackRequest:
      type: object
      properties:
        feedback:
          type: string
          enum: [ACCEPTED, REJECTED]
      required: [feedback]
    VulnerabilityFeedback:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [ACCEPTED, REJECTED]
        userFeedback:
          type: string
          enum: [ACCEPTED, REJECTED]
      required: [id, status, userFeedback]
    ReportRequestAccepted:
      type: object
      properties:
        reportId:
          type: string
        status:
          type: string
          enum: [GENERATING, READY]
        message:
          type: string
      required: [reportId, status, message]
    AuthUserResponse:
      $ref: '#/components/schemas/ApiEnvelopeAuthUser'
    ConnectedRepoListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/ConnectedRepoList'
        message:
          type: [string, 'null']
        timestamp:
          type: string
      required: [success, data, message, timestamp]
    ConnectedRepoResponse:
      $ref: '#/components/schemas/ApiEnvelopeConnectedRepo'
    ScanRequestResponse:
      $ref: '#/components/schemas/ApiEnvelopeScanRequest'
    ScanDetailResponse:
      $ref: '#/components/schemas/ApiEnvelopeScanDetail'
    RepoScanPageResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/RepoScanPage'
        message:
          type: [string, 'null']
        timestamp:
          type: string
      required: [success, data, message, timestamp]
    VulnerabilityDetailResponse:
      $ref: '#/components/schemas/ApiEnvelopeVulnerabilityDetail'
    VulnerabilityFeedbackResponse:
      $ref: '#/components/schemas/ApiEnvelopeVulnerabilityFeedback'
    DashboardResponse:
      $ref: '#/components/schemas/ApiEnvelopeDashboard'
    ReportDetailResponse:
      $ref: '#/components/schemas/ApiEnvelopeReportDetail'
    ReportRequestResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/ReportRequestAccepted'
        message:
          type: [string, 'null']
        timestamp:
          type: string
      required: [success, data, message, timestamp]